requirements.txt
alembic==1.3.3
aniso8601==7.0.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-GraphQL==2.0.1
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.1
graphql-server-core==1.1.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.1
MarkupSafe==1.1.1
promise==2.3
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==0.16.0
